refactor(sphere): remove the L1/ALPHA layer from the frontend
Delete the L1 wallet UI (components, modals, hooks, MineAlpha page,
L1 balance display) and rewire SDK integration, onboarding/shared,
the L3 panel, connect, and static page copy to the L3-only model.
Drop the dead alpha1-address branch in chat identifier normalization.
