+++ src/components/layout/DashboardLayout.tsx
@@ -12,7 +12,6 @@ import bgPosterUrl from '/bg-poster.webp';
 
 export function DashboardLayout() {
   const location = useLocation();
-  const isMinePage = location.pathname === '/mine';
   const { isFullscreen } = useUIState();
   const { activeTabId } = useDesktopState();
   const tutorial = useTutorial();
@@ -59,9 +58,7 @@ export function DashboardLayout() {
       <div className="relative z-10 flex flex-col h-full">
         {!isFullscreen && <Header />}
         <div className={`flex-1 min-h-0 flex ${!isAgentPage ? 'overflow-y-auto overflow-x-hidden' : ''}`}>
-          <div className={`flex-1 w-full ${isFullscreen ? 'p-0' : isAgentPage ? 'px-0 sm:px-12 lg:px-28 pb-0' : 'px-0 sm:px-12 lg:px-28 pt-4 pb-0 md:pt-8 lg:pb-8'} ${
-            isMinePage ? 'bg-neutral-100 dark:bg-transparent' : ''
-          }`}>
+          <div className={`flex-1 w-full ${isFullscreen ? 'p-0' : isAgentPage ? 'px-0 sm:px-12 lg:px-28 pb-0' : 'px-0 sm:px-12 lg:px-28 pt-4 pb-0 md:pt-8 lg:pb-8'}`}>
             <Outlet />
           </div>
         </div>
