--- src/components/wallet/L1/components/modals/index.ts
@@ -1,7 +1,0 @@
-export { QRModal } from "./QRModal";
-export { SaveWalletModal } from "./SaveWalletModal";
-export { LoadPasswordModal } from "./LoadPasswordModal";
-export { DeleteConfirmationModal } from "./DeleteConfirmationModal";
-export { TransactionConfirmationModal } from "./TransactionConfirmationModal";
-export { BridgeModal } from "./BridgeModal";
-export { SendModal } from "./SendModal";
