--- src/components/wallet/L1/hooks/index.ts
@@ -1,1 +1,0 @@
-// L1 hooks barrel - legacy hooks removed, see SDK hooks in src/sdk/hooks/
