--- src/components/wallet/L1/modals/index.ts
@@ -1,1 +1,0 @@
-export { L1WalletModal } from "./L1WalletModal";
