--- src/components/wallet/L1/views/index.ts
@@ -1,1 +1,0 @@
-// L1 views barrel - legacy views removed, see L1WalletModal for current L1 UI
