+++ src/components/wallet/L3/views/L3WalletView.tsx
@@ -1,9 +1,9 @@
-import { Plus, ArrowUpRight, ArrowDownUp, Loader2, Coins, Layers, Eye, EyeOff, Wifi } from 'lucide-react';
+import { Plus, ArrowUpRight, ArrowDownUp, Loader2, Coins, Layers, Eye, EyeOff, Wifi, Shield, AlertCircle, FileJson } from 'lucide-react';
 import { AnimatePresence, motion, useMotionValue, useTransform, animate } from 'framer-motion';
 import { AssetRow } from '../../shared/components';
 import { useEffect, useMemo, useRef, useState, useCallback } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useIdentity, useAssets, useTokens, useL1Balance } from '../../../../sdk';
+import { useIdentity, useAssets, useTokens } from '../../../../sdk';
 import { useSphereContext } from '../../../../sdk/hooks/core/useSphere';
 import { CreateWalletFlow } from '../../onboarding/CreateWalletFlow';
 import { TokenRow } from '../../shared/components';
@@ -15,7 +15,8 @@ import { SeedPhraseModal } from '../modals/SeedPhraseModal';
 import { TransactionHistoryModal } from '../modals/TransactionHistoryModal';
 import { SettingsModal } from '../modals/SettingsModal';
 import { BackupWalletModal, LogoutConfirmModal } from '../../shared/modals';
-import { SaveWalletModal } from '../../L1/components/modals';
+import { WalletScreen } from '../../ui/WalletScreen';
+import { ModalHeader } from '../../ui';
 type Tab = 'assets' | 'tokens';
 
 // Animated balance display with smooth number transitions
@@ -75,17 +76,14 @@ function BalanceDisplay({
 // Inline status line showing current wallet activity
 function WalletStatusLine({
   isLoadingAssets,
-  isLoadingL1,
   pendingCount,
 }: {
   isLoadingAssets: boolean;
-  isLoadingL1: boolean;
   pendingCount: number;
 }) {
   const items: { label: string; spinning?: boolean }[] = [];
 
   if (isLoadingAssets) items.push({ label: 'Loading assets', spinning: true });
-  if (isLoadingL1) items.push({ label: 'Loading L1 balance', spinning: true });
   if (pendingCount > 0) items.push({ label: `${pendingCount} pending transfer${pendingCount > 1 ? 's' : ''}` });
 
   if (items.length === 0) return null;
@@ -122,8 +120,6 @@ interface L3WalletViewProps {
   setIsRequestsOpen: (value: boolean) => void;
   isSettingsOpen: boolean;
   setIsSettingsOpen: (value: boolean) => void;
-  isL1WalletOpen: boolean;
-  setIsL1WalletOpen: (value: boolean) => void;
   paymentRequests: import('../hooks/useIncomingPaymentRequests').IncomingPaymentRequest[];
   paymentRequestsPendingCount: number;
   paymentRequestsReject: (request: import('../hooks/useIncomingPaymentRequests').IncomingPaymentRequest) => Promise<void>;
@@ -140,7 +136,6 @@ export function L3WalletView({
   setIsRequestsOpen,
   isSettingsOpen,
   setIsSettingsOpen,
-  setIsL1WalletOpen,
   paymentRequests,
   paymentRequestsPendingCount,
   paymentRequestsReject,
@@ -153,19 +148,12 @@ export function L3WalletView({
   const { identity, isLoading: isLoadingIdentity } = useIdentity();
   const { assets: sdkAssets, isLoading: isLoadingAssets } = useAssets();
   const { tokens: sdkTokens, pendingTokens } = useTokens();
-  const { balance: l1BalanceData, isLoading: isLoadingL1 } = useL1Balance();
   const { sphere, deleteWallet } = useSphereContext();
 
   const assets = sdkAssets;
 
   const tokens = sdkTokens;
   const sendableTokens = useMemo(() => tokens.filter(t => t.coinId !== 'NAMETAG'), [tokens]);
-
-  // L1 balance as a number (ALPHA units)
-  const l1Balance = useMemo(() => {
-    if (!l1BalanceData?.total) return 0;
-    return Number(l1BalanceData.total) / 1e8;
-  }, [l1BalanceData]);
 
   const [activeTab, setActiveTab] = useState<Tab>('assets');
   const [isSendModalOpen, setIsSendModalOpen] = useState(false);
@@ -201,16 +189,13 @@ export function L3WalletView({
     }
 
     const newIds = new Set<string>();
-    if (l1Balance > 0 && !prevAssetCoinIdsRef.current.has('l1-alpha')) {
-      newIds.add('l1-alpha');
-    }
     assets.forEach(asset => {
       if (!prevAssetCoinIdsRef.current.has(asset.coinId)) {
         newIds.add(asset.coinId);
       }
     });
     return newIds;
-  }, [assets, l1Balance]);
+  }, [assets]);
 
   // Update previous snapshots after render (for next comparison)
   useEffect(() => {
@@ -221,9 +206,8 @@ export function L3WalletView({
 
   useEffect(() => {
     const currentIds = new Set(assets.map(a => a.coinId));
-    if (l1Balance > 0) currentIds.add('l1-alpha');
     prevAssetCoinIdsRef.current = currentIds;
-  }, [assets, l1Balance]);
+  }, [assets]);
 
   // New modal states
   const [isBackupOpen, setIsBackupOpen] = useState(false);
@@ -235,39 +219,10 @@ export function L3WalletView({
     setShowBalances(!showBalances);
   }, [showBalances, setShowBalances]);
 
-  // Create L1 ALPHA asset
-  const l1AlphaAsset = useMemo(() => {
-    const satoshis = BigInt(Math.round(l1Balance * 100000000));
-    const totalAmount = satoshis.toString();
-    const fiatValue = l1Balance * 1.0; // priceUsd = 1.0
-    return {
-      coinId: 'l1-alpha',
-      symbol: 'ALPHA',
-      name: 'Unicity Alphanet',
-      totalAmount,
-      decimals: 8,
-      tokenCount: 1,
-      confirmedAmount: totalAmount,
-      unconfirmedAmount: '0',
-      transferringAmount: '0',
-      confirmedTokenCount: 1,
-      unconfirmedTokenCount: 0,
-      transferringTokenCount: 0,
-      iconUrl: undefined,
-      priceUsd: 1.0,
-      priceEur: 0.92,
-      change24h: 0,
-      fiatValueUsd: fiatValue,
-      fiatValueEur: fiatValue * 0.92,
-    } satisfies import('@unicitylabs/sphere-sdk').Asset;
-  }, [l1Balance]);
-
   const totalValue = useMemo(() => {
     // Sum up L3 asset values (using SDK-provided fiat values for accuracy)
-    const l3Value = sdkAssets.reduce((sum, asset) => sum + (asset.fiatValueUsd ?? 0), 0);
-    const l1Value = l1AlphaAsset.fiatValueUsd ?? 0;
-    return l3Value + l1Value;
-  }, [sdkAssets, l1AlphaAsset]);
+    return sdkAssets.reduce((sum, asset) => sum + (asset.fiatValueUsd ?? 0), 0);
+  }, [sdkAssets]);
 
   const handleShowSeedPhrase = () => {
     if (!sphere) return;
@@ -329,18 +284,12 @@ export function L3WalletView({
     setIsBackupOpen(true);
   };
 
-  // Format L1 balance for settings modal
-  const formatL1Balance = (balance: number) => {
-    return balance.toLocaleString('en-US', { minimumFractionDigits: 4, maximumFractionDigits: 4 });
-  };
-
   if (isLoadingIdentity) {
     return (
       <div className="flex flex-col items-center justify-center h-full gap-3">
         <Loader2 className="animate-spin text-neutral-400 dark:text-neutral-600" />
         <WalletStatusLine
           isLoadingAssets={isLoadingAssets}
-          isLoadingL1={isLoadingL1}
           pendingCount={0}
         />
       </div>
@@ -364,7 +313,6 @@ export function L3WalletView({
           />
           <WalletStatusLine
             isLoadingAssets={isLoadingAssets}
-            isLoadingL1={isLoadingL1}
             pendingCount={pendingTokens.length}
           />
         </div>
@@ -444,34 +392,19 @@ export function L3WalletView({
               {/* ASSETS VIEW - no container animation, only item animations */}
               {activeTab === 'assets' && (
                 <div className="space-y-2">
-                  {assets.length === 0 && l1Balance === 0 ? (
+                  {assets.length === 0 ? (
                     <EmptyState />
                   ) : (
-                    <>
-                      {/* L1 ALPHA - only show if balance > 0 */}
-                      {l1Balance > 0 && (
-                        <AssetRow
-                          key="l1-alpha"
-                          asset={l1AlphaAsset}
-                          showBalances={showBalances}
-                          delay={newAssetCoinIds.has('l1-alpha') ? 0 : 0}
-                          layer="L1"
-                          isNew={newAssetCoinIds.has('l1-alpha')}
-                          onClick={() => setIsL1WalletOpen(true)}
-                        />
-                      )}
-                      {/* L3 Assets */}
-                      {assets.map((asset, index) => (
-                        <AssetRow
-                          key={asset.coinId}
-                          asset={asset}
-                          showBalances={showBalances}
-                          delay={newAssetCoinIds.has(asset.coinId) ? (index + 1) * 0.05 : 0}
-                          layer="L3"
-                          isNew={newAssetCoinIds.has(asset.coinId)}
-                        />
-                      ))}
-                    </>
+                    assets.map((asset, index) => (
+                      <AssetRow
+                        key={asset.coinId}
+                        asset={asset}
+                        showBalances={showBalances}
+                        delay={newAssetCoinIds.has(asset.coinId) ? (index + 1) * 0.05 : 0}
+                        layer="L3"
+                        isNew={newAssetCoinIds.has(asset.coinId)}
+                      />
+                    ))
                   )}
                 </div>
               )}
@@ -525,10 +458,8 @@ export function L3WalletView({
       <SettingsModal
         isOpen={isSettingsOpen}
         onClose={() => setIsSettingsOpen(false)}
-        onOpenL1Wallet={() => setIsL1WalletOpen(true)}
         onBackupWallet={() => setIsBackupOpen(true)}
         onLogout={() => setIsLogoutConfirmOpen(true)}
-        l1Balance={formatL1Balance(l1Balance)}
         hasMnemonic={hasMnemonic}
       />
 
@@ -571,5 +502,146 @@ function EmptyState({ text }: { text?: string }) {
         {text || <>Wallet is empty.<br />Mint some tokens to start!</>}
       </div>
     </div>
+  );
+}
+
+// Wallet-export ("Backup Wallet") modal. Generic L3 wallet JSON export —
+// inlined here because it is consumed only by this view.
+interface SaveWalletModalProps {
+  show: boolean;
+  onConfirm: (filename: string, password?: string) => void;
+  onCancel: () => void;
+  /** Whether mnemonic is available (shows indicator) */
+  hasMnemonic?: boolean;
+  /** Default filename (without extension) */
+  defaultFilename?: string;
+}
+
+function SaveWalletModal({ show, onConfirm, onCancel, hasMnemonic, defaultFilename = "sphere_wallet_backup" }: SaveWalletModalProps) {
+  const [filename, setFilename] = useState(defaultFilename);
+  const [password, setPassword] = useState("");
+  const [passwordConfirm, setPasswordConfirm] = useState("");
+  const [error, setError] = useState("");
+
+  const handleConfirm = () => {
+    setError("");
+    if (password) {
+      if (password !== passwordConfirm) {
+        setError("Passwords do not match!");
+        return;
+      }
+      if (password.length < 4) {
+        setError("Password must be at least 4 characters");
+        return;
+      }
+    }
+
+    onConfirm(filename, password || undefined);
+
+    // Reset state
+    setFilename(defaultFilename);
+    setPassword("");
+    setPasswordConfirm("");
+    setError("");
+  };
+
+  return (
+    <WalletScreen isOpen={show} onClose={onCancel}>
+      <ModalHeader variant="screen" title="Backup Wallet" onClose={onCancel} />
+      <div className="px-6 py-8 flex flex-col flex-1 overflow-y-auto">
+        <div className="flex flex-col items-center text-center mb-6">
+          <motion.div
+            initial={{ scale: 0 }}
+            animate={{ scale: 1 }}
+            transition={{ delay: 0.1, type: "spring", stiffness: 200 }}
+            className="w-16 h-16 bg-orange-500/10 rounded-full flex items-center justify-center mb-4"
+          >
+            <motion.div
+              animate={{ y: [0, -2, 0] }}
+              transition={{ delay: 0.2, duration: 1.5, repeat: Infinity, ease: "easeInOut" }}
+            >
+              <Shield className="w-7 h-7 text-orange-500" />
+            </motion.div>
+          </motion.div>
+          <p className="text-xs text-neutral-500 dark:text-white/45">
+            Export your wallet keys to a JSON file. Keep this safe!
+          </p>
+        </div>
+
+        {/* Format indicator */}
+        <div className="flex items-center justify-center gap-2 mb-4 p-2 bg-orange-500/10 border border-orange-500/20 rounded-lg">
+          <FileJson className="w-4 h-4 text-orange-500" />
+          <span className="text-sm text-orange-500 font-medium">JSON Format</span>
+          {hasMnemonic && (
+            <span className="text-[10px] bg-green-500/20 text-green-500 px-1.5 py-0.5 rounded">
+              +mnemonic
+            </span>
+          )}
+        </div>
+
+        <p className="text-[10px] text-neutral-400 mb-4 text-center">
+          Includes verification address{hasMnemonic ? " and recovery phrase" : ""}
+        </p>
+
+        <label className="text-xs text-neutral-500 mb-1 block">
+          Filename
+        </label>
+        <input
+          placeholder="Filename"
+          value={filename}
+          onChange={(e) => setFilename(e.target.value)}
+          className="w-full mb-3 px-4 py-3 bg-neutral-100 dark:bg-white/6 rounded-xl text-neutral-800 dark:text-neutral-200 placeholder-neutral-400 border border-neutral-200 dark:border-white/8 focus:border-orange-500 outline-none transition-colors"
+        />
+
+        <label className="text-xs text-neutral-500 mb-1 block">
+          Encryption Password (Optional)
+        </label>
+        <input
+          placeholder="Password"
+          type="password"
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+          className="w-full mb-3 px-4 py-3 bg-neutral-100 dark:bg-white/6 rounded-xl text-neutral-800 dark:text-neutral-200 placeholder-neutral-400 border border-neutral-200 dark:border-white/8 focus:border-orange-500 outline-none transition-colors"
+        />
+
+        <input
+          placeholder="Confirm Password"
+          type="password"
+          value={passwordConfirm}
+          onChange={(e) => setPasswordConfirm(e.target.value)}
+          className="w-full mb-4 px-4 py-3 bg-neutral-100 dark:bg-white/6 rounded-xl text-neutral-800 dark:text-neutral-200 placeholder-neutral-400 border border-neutral-200 dark:border-white/8 focus:border-orange-500 outline-none transition-colors"
+        />
+
+        {error && (
+          <motion.div
+            initial={{ opacity: 0, y: -10 }}
+            animate={{ opacity: 1, y: 0 }}
+            className="mb-4 p-3 bg-red-500/10 border border-red-900/50 rounded-lg flex items-center gap-2"
+          >
+            <AlertCircle className="w-4 h-4 text-red-500 shrink-0" />
+            <span className="text-red-400 text-sm">{error}</span>
+          </motion.div>
+        )}
+
+        <div className="flex gap-3 mt-2">
+          <motion.button
+            whileHover={{ scale: 1.02 }}
+            whileTap={{ scale: 0.98 }}
+            onClick={onCancel}
+            className="flex-1 py-3 bg-neutral-100 dark:bg-white/6 rounded-xl text-neutral-700 dark:text-white/65 hover:bg-neutral-200 dark:hover:bg-white/10 transition-colors font-medium"
+          >
+            Cancel
+          </motion.button>
+          <motion.button
+            whileHover={{ scale: 1.02 }}
+            whileTap={{ scale: 0.98 }}
+            onClick={handleConfirm}
+            className="flex-1 py-3 bg-orange-500 hover:bg-orange-600 rounded-xl text-white font-medium transition-all shadow-lg shadow-orange-500/20"
+          >
+            Save
+          </motion.button>
+        </div>
+      </div>
+    </WalletScreen>
   );
 }
