+++ src/components/wallet/shared/components/AssetRow.tsx
@@ -8,7 +8,7 @@ interface AssetRowProps {
   showBalances: boolean;
   delay: number;
   onClick?: () => void;
-  layer?: 'L1' | 'L3';
+  layer?: 'L3';
   /** If true, animate entrance. If false, render without animation (asset was already shown) */
   isNew?: boolean;
 }
@@ -114,11 +114,7 @@ export const AssetRow = memo(function AssetRow({ asset, showBalances, delay, onC
           <div className="flex items-center gap-2">
             <div className="text-neutral-900 dark:text-[#fefefe] font-medium text-sm" style={{ fontFamily: "'Geist Mono', 'SF Mono', 'Fira Code', monospace" }}>{asset.symbol}</div>
             {layer && (
-              <span className={`px-1.5 py-0.5 text-[10px] font-bold rounded ${
-                layer === 'L1'
-                  ? 'bg-blue-500/20 text-blue-600 dark:text-blue-400'
-                  : 'bg-orange-500/20 text-orange-600 dark:text-brand-orange'
-              }`}>
+              <span className="px-1.5 py-0.5 text-[10px] font-bold rounded bg-orange-500/20 text-orange-600 dark:text-brand-orange">
                 {layer}
               </span>
             )}
