+++ src/sdk/hooks/index.ts
@@ -28,19 +28,6 @@ export type { UseTopUpReturn, TopUpResult } from './payments/useTopUp';
 export { useTransactionHistory } from './payments/useTransactionHistory';
 export type { UseTransactionHistoryReturn } from './payments/useTransactionHistory';
 
-// L1
-export { useL1Balance } from './l1/useL1Balance';
-export type { UseL1BalanceReturn, L1BalanceData } from './l1/useL1Balance';
-export { useL1Utxos } from './l1/useL1Utxos';
-export type { UseL1UtxosReturn, Utxo } from './l1/useL1Utxos';
-export { useL1Send } from './l1/useL1Send';
-export type { UseL1SendReturn, L1SendParams, L1SendResult } from './l1/useL1Send';
-export { useL1Transactions } from './l1/useL1Transactions';
-export type {
-  UseL1TransactionsReturn,
-  L1Transaction,
-} from './l1/useL1Transactions';
-
 // Communications
 export { useSendDM } from './comms/useSendDM';
 export type { UseSendDMReturn } from './comms/useSendDM';
